Rename model_exec_request_handler file.
Renaming nwm_request_handler.py to model_exec_request_handler.py after
changing module so that the class is now ModelExecRequestHandler and
responsible for all ModelExecRequest types.
