+++ python/lib/externalrequests/dmod/externalrequests/__init__.py
@@ -1,5 +1,5 @@
 from .auth_handler import AuthHandler
 from .maas_request_handlers import MaaSRequestHandler, PartitionRequestHandler
-from .nwm_request_handler import ModelExecRequestHandler
+from .model_exec_request_handler import ModelExecRequestHandler
 
 name = 'externalrequests'
